app/cmd/requirements.txt
# Packages requirements for updating the requirements.txt.freeze list via
#  make update-packages
# Deployments will use requirements.txt.freeze which is generated by make update-packages

# Utils
python-dotenv
sqlalchemy
pydantic
bcrypt
pendulum
psycopg2-binary
requests
pytz
passlib

# Messaging
ansq
gnsq

# WebServer
fastapi
fastapi-login
uvicorn
loguru
jinja2
orjson
structlog
asgi_correlation_id
user_agents
dataclasses_json
authlib

# Integration

# Testing
pytest
pytest-cov
pytest-mock